Polar/Dockerfile/requirements.txt
absl-py==1.4.0
batchgenerators==0.25
blobfile==2.0.2
brotlipy==0.7.0
cachetools==5.3.1
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
fonttools==4.42.0
future==0.18.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
imageio==2.31.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
lazy_loader==0.3
lightning-utilities==0.9.0
linecache2==1.0.0
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
monai==1.2.0
networkx==3.1
nibabel==5.1.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.4.0
protobuf==4.24.0
psutil==5.9.5
pyasn1==0.5.0
opencv-python-headless==4.8.1.78
pyasn1-modules==0.3.0
pycryptodomex==3.18.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
SimpleITK==2.2.1
six==1.16.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tifffile==2023.8.12
torch-geometric==2.3.1
torchmetrics==1.0.3
torchtyping==0.1.4
tqdm==4.66.1
traceback2==1.4.0
typeguard==4.1.1
unittest2==1.1.0
Werkzeug==2.3.7
